combine benser-cli and browser into wgpu-renderer crate
