--- benser-cli/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "benser-cli"
-version = "0.1.0"
-edition = "2021"
-
-
-[dependencies]
-benser = { path = "../benser" }
-html = { path = "../html" }
-paint = { path = "../paint" }
-clap = { version = "4.0.18", features = ["derive"] }
-image = "0.24.4"
-env_logger = "0.10.0"
-log = "0.4.18"
-pollster = "0.3.0"
-wgpu = "0.16.1"
-wgpu_text = "0.8.1"
-futures-intrusive = "0.5.0"
-bytemuck = { version = "1.13.1", features = ["derive"] }
-lyon = "1.0.1"
-
