--- benser-cli/src/lib.rs
@@ -1,51 +1,0 @@
-use std::path::PathBuf;
-
-use clap::Parser;
-
-#[derive(Parser, Debug)]
-pub struct Args {
-    /// CSS file to use
-    pub css_file: PathBuf,
-
-    /// HTML file to use
-    pub html_file: PathBuf,
-
-    /// File to output to
-    pub output: PathBuf,
-
-    /// Viewport width
-    #[arg(long = "width")]
-    pub viewport_width: Option<f32>,
-
-    /// Viewport height
-    #[arg(long = "height")]
-    pub viewport_height: Option<f32>,
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
-pub struct Vertex {
-    pub position: [f32; 3],
-    pub color: [f32; 3],
-}
-
-impl Vertex {
-    pub fn desc() -> wgpu::VertexBufferLayout<'static> {
-        wgpu::VertexBufferLayout {
-            array_stride: std::mem::size_of::<Vertex>() as wgpu::BufferAddress,
-            step_mode: wgpu::VertexStepMode::Vertex,
-            attributes: &[
-                wgpu::VertexAttribute {
-                    offset: 0,
-                    shader_location: 0,
-                    format: wgpu::VertexFormat::Float32x3,
-                },
-                wgpu::VertexAttribute {
-                    offset: std::mem::size_of::<[f32; 3]>() as wgpu::BufferAddress,
-                    shader_location: 1,
-                    format: wgpu::VertexFormat::Float32x3,
-                },
-            ],
-        }
-    }
-}
