+++ wgpu-renderer/examples/rainbow/rainbow.html
