+++ wgpu-renderer/src/main.rs
@@ -1,72 +1,13 @@
-use wgpu_renderer::State;
-use wgpu_text::section::{Section, Text};
-use winit::{
-    event::*,
-    event_loop::{ControlFlow, EventLoop},
-    window::WindowBuilder,
-};
+use clap::Parser;
+use wgpu_renderer::{args::Args, browser, file_output};
 
 fn main() {
-    pollster::block_on(run());
-}
-
-async fn run() {
     env_logger::init();
-    let event_loop = EventLoop::new();
-    let window = WindowBuilder::new()
-        .with_title("benser")
-        .build(&event_loop)
-        .unwrap();
+    let args = Args::parse();
 
-    let mut state = State::new(window).await;
-
-    // Directly implemented from glyph_brush.
-    let section = Section::default().add_text(Text::new("Hello World"));
-
-    event_loop.run(move |event, _, control_flow| match event {
-        Event::RedrawRequested(window_id) if window_id == state.window().id() => {
-            state.update();
-            match state.render() {
-                Ok(_) => {}
-                // Reconfigure the surface if lost
-                Err(wgpu::SurfaceError::Lost) => state.resize(state.size),
-                // The system is out of memory, we should probably quit
-                Err(wgpu::SurfaceError::OutOfMemory) => *control_flow = ControlFlow::Exit,
-                // All other errors (Outdated, Timeout) should be resolved by the next frame
-                Err(e) => eprintln!("{:?}", e),
-            }
-        }
-        Event::MainEventsCleared => {
-            // RedrawRequested will only trigger once, unless we manually
-            // request it.
-            state.window().request_redraw();
-        }
-        Event::WindowEvent {
-            ref event,
-            window_id,
-        } if window_id == state.window().id() => {
-            if !state.input(event) {
-                match event {
-                    WindowEvent::Resized(physical_size) => {
-                        state.resize(*physical_size);
-                    }
-                    WindowEvent::ScaleFactorChanged { new_inner_size, .. } => {
-                        state.resize(**new_inner_size);
-                    }
-                    WindowEvent::CloseRequested
-                    | WindowEvent::KeyboardInput {
-                        input:
-                            KeyboardInput {
-                                state: ElementState::Pressed,
-                                virtual_keycode: Some(VirtualKeyCode::Escape),
-                                ..
-                            },
-                        ..
-                    } => *control_flow = ControlFlow::Exit,
-                    _ => {}
-                }
-            }
-        }
-        _ => {}
-    });
+    if let Some(ref path) = args.output {
+        pollster::block_on(file_output::run(args))
+    } else {
+        pollster::block_on(browser::run(args));
+    }
 }
